cleaning up the services and adding zip service
