+++ src/electionguard-ui/ElectionGuard.UI.Lib/Models/FileContent.cs
@@ -1,0 +1,6 @@
+﻿namespace ElectionGuard.UI.Lib.Models;
+
+/// <summary>
+/// structure to hold data for files to write to zip file
+/// </summary>
+public record FileContents(string FileName, string Contents);
