+++ src/electionguard-ui/ElectionGuard.UI.Lib/Services/Database/BallotService.cs
