+++ src/electionguard-ui/ElectionGuard.UI.Lib/Services/Database/ManifestService.cs
