+++ src/electionguard-ui/ElectionGuard.UI.Lib/Services/Database/PlaintextSpoiledBallotService.cs
