+++ src/electionguard-ui/ElectionGuard.UI.Lib/Services/Database/PlaintextTallyService.cs
