+++ src/electionguard-ui/ElectionGuard.UI.Lib/Services/Database/TallyJoinedService.cs
