+++ src/electionguard-ui/ElectionGuard.UI.Lib/Services/Database/TallyService.cs
